pyproject.toml
[project]
name = "aie6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
]
